pyproject.toml
[tool.poetry]
name = "homework_10_1"
version = "0.1.0"
description = ""
authors = ["Andrey Kazantsev <andrey@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.1"
requests = "^2.32.4"

[tool.pytest.ini_options]
addopts = "--cov=scr --cov-report=html"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"


[tool.poetry.group.lit.dependencies]
pytest = "^8.3.2"
mypy = "^1.11.2"
isort = "^5.13.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119

[tool.isort]
line_length = 119


